Move server code to package
