+++ server/lambda.go
@@ -1,8 +1,8 @@
 /*
- * Copyright 2023 Hypermode, Inc.
+ * Copyright 2024 Hypermode, Inc.
  */
 
-package main
+package server
 
 import (
 	"context"
@@ -11,12 +11,9 @@ import (
 	"net/http"
 	"strings"
 
-	"hmruntime/config"
 	"hmruntime/functions"
 	"hmruntime/host"
 	"hmruntime/logger"
-
-	"github.com/rs/xid"
 )
 
 type HMRequest struct {
@@ -38,11 +35,7 @@ type HMError struct {
 	Message string `json:"message"`
 }
 
-type AdminRequest struct {
-	Action string `json:"action"`
-}
-
-func handleRequest(w http.ResponseWriter, r *http.Request) {
+func handleDgraphLambdaRequest(w http.ResponseWriter, r *http.Request) {
 
 	// Assign an Execution ID to the request context
 	ctx, r := assignExecutionId(w, r)
@@ -222,52 +215,4 @@ func writeDataAsJson(w http.ResponseWriter, data any, isJson bool) error {
 	}
 
 	return nil
-}
-
-func handleAdminRequest(w http.ResponseWriter, r *http.Request) {
-
-	// Assign an Execution ID to the request context
-	ctx, r := assignExecutionId(w, r)
-
-	// Decode the request body
-	var req AdminRequest
-	dec := json.NewDecoder(r.Body)
-	err := dec.Decode(&req)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		logger.Err(ctx, err).Msg("Failed to decode request body.")
-		return
-	}
-
-	// Perform the requested action
-	switch req.Action {
-	// TODO: Add admin actions here
-	default:
-		err = fmt.Errorf("unknown action: %s", req.Action)
-	}
-
-	// Write the response
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		logger.Err(ctx, err).Msg("Failed to perform admin action.")
-	} else {
-		w.WriteHeader(http.StatusOK)
-	}
-}
-
-func startServer(ctx context.Context) error {
-	logger.Info(ctx).
-		Int("port", config.Port).
-		Msg("Listening for incoming requests.")
-	http.HandleFunc("/graphql-worker", handleRequest)
-	http.HandleFunc("/admin", handleAdminRequest)
-	return http.ListenAndServe(fmt.Sprintf(":%d", config.Port), nil)
-}
-
-func assignExecutionId(w http.ResponseWriter, r *http.Request) (context.Context, *http.Request) {
-	executionId := xid.New().String()
-	w.Header().Add("X-Hypermode-ExecutionID", executionId)
-
-	ctx := context.WithValue(r.Context(), logger.ExecutionIdContextKey, executionId)
-	return ctx, r.WithContext(ctx)
 }
