+++ server/server.go
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2024 Hypermode, Inc.
+ */
+
+package server
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+
+	"hmruntime/config"
+	"hmruntime/logger"
+
+	"github.com/rs/xid"
+)
+
+func Start(ctx context.Context) error {
+	logger.Info(ctx).
+		Int("port", config.Port).
+		Msg("Listening for incoming requests.")
+	http.HandleFunc("/graphql-worker", handleDgraphLambdaRequest)
+	http.HandleFunc("/admin", handleAdminRequest)
+	return http.ListenAndServe(fmt.Sprintf(":%d", config.Port), nil)
+}
+
+func assignExecutionId(w http.ResponseWriter, r *http.Request) (context.Context, *http.Request) {
+	executionId := xid.New().String()
+	w.Header().Add("X-Hypermode-ExecutionID", executionId)
+
+	ctx := context.WithValue(r.Context(), logger.ExecutionIdContextKey, executionId)
+	return ctx, r.WithContext(ctx)
+}
